ApiAssistent/requirements.txt
openai-whisper
openai
sounddevice
numpy
pyttsx3